fixed PDF stuff
lots of small changes in the text

ONLY COMPILES WITH PDFLaTeX  (for proper generation using png and jpg
images; the postscript versions are truly ugly).  SORRY, BUT IT NEEDS
TO BE THIS WAY FOR GENERATION OF A REASONABLE COPY OF THE IMAGES.

One more batch of changes to make, and then I need people to take one
last look at it.  At this point, I'll provide the PDF file if people
can't compile it locally.


git-svn-id: https://svn.r-project.org/ESS/trunk@1689 0bbaf3bd-34e0-0310-bf65-c717079852d4