ess-mous is ready for testing.
ess-mous.el: put ess-mous on submenu of C-mouse-3 for
             ess-transript-mode, inferior-ess-mode, ess-mode
             This feature is still beta.

ess-inf.el:
  ess-ddeclient-p. now depends on ess-local-process-name
  get-ess-process. I changed error message to say buffer is not
                   associated with an ESS process (instead of
                   saying that no ESS process is running)
  inferior-ess-mode-menu.  Added "What is this? (beta)"

ess-trns.el:
  ess-transcript-mode-menu.  Added "What is this? (beta)"

ess-mode.el:
  ess-mode-menu.  Added "What is this? (beta)"

ChangeLog: ess-mous changes.

I tested it on windows for R, S+4, and Sqpe+4.
I am having line-ending problems with Sqpe+4 but otherwise it looks ok.
Please look at it.


I found a problem which I am unsure how to solve.  The variable
`ess-current-process-name' in the `*R*' buffer can have the value
"S+4".  All I have to do is open an S+4 process and an R process, then
C-c C-s the file.s buffer to S+4 and then switch to the R process.

I think, but am willing to hear arguments against it, that anytime an
inferior-ess-mode buffer is made current, then it should automatically
reset `ess-current-process-name' to refer to itself.  I'm not sure how
to do that.  This is explicitly not a buffer-local variable.  Is there
a hook that gets called when we change into a buffer, either by C-x b
or by clicking?


git-svn-id: https://svn.r-project.org/ESS/trunk@1920 0bbaf3bd-34e0-0310-bf65-c717079852d4