I finally read the latest essa-sas.el
1. I found ess-sas-toggle-sas-mode several months ago and started to use it.
I like it.  Now that I have read it closely I revised it and renamed it.
;;; this version of ess-sas-toggle-log-sas-mode
;;; 1. can be included in .emacs as
;;;           (ess-sas-toggle-log-sas-mode t)
;;;    because it doesn't need to find a .log file.
;;; 2. works if there is already an association for .log files.
;;; 3. restores the old association when SAS-mode is deleted.
;;; 4. works with the file it has, doesn't kill it and get a fresh copy.
;;;    This matters for large log files and slow telephone connections.
;;; 5. I changed its name to be more descriptive.


2. How is kermit used by ESS?

I see definititions for ess-kermit-command ess-kermit-get ess-kermit-send,
but do not see any places they are used.

I think these three items should be separated into a new ess-kermit.el
rather than buried in essa-sas.el.  Are they restricted to ess in any way?
Might they be better as a new kermit.el?


3. You hardwired the `txt' extension into ess-sas-file-path.  That should
be sensitive to the value of ess-sas-suffix-1 and ess-sas-suffix-2.


4. ess-sas-submit-mac is the only ess-sas-submit* that does not have
ess-sas-submit-command-options.  I don't understand
ess-sas-submit-command-options.  Why is it buffer-local?  I think it
should depend on the host and the version of sas, not on the
myfile.sas command file.  Am I misunderstanding what is going on?


5. ess-revert-wisely in essa-sas.el and ess-revert in essl-bug.el are identical.
This suggests that the ess-revert name should be used and moved to ess-utils.el


git-svn-id: https://svn.r-project.org/ESS/trunk@1921 0bbaf3bd-34e0-0310-bf65-c717079852d4