1. what is going on?  I write    \newcommand*{\regstrd}{$^{\mbox{\scriptsize\textregistered}}$} someone else changes it to    \newcommand*{\regstrd}{$^{\mbox{\scriptsize{\textregistered}}}$}
Since \scriptsize is a declaration, not a function, what is the
intent of the extra braces?  TeX sees the keyword \textregistered
as a single token, it does not see it as a string of characters.


2. I wrote Interactive Processing/S by taking a few paragraphs from
the previous version in doc/techrep/ess.tex and editing them.


3. I wrote a new S Batch section.


4. I think of the unnumbered section {ESS Resources on the Internet}
as structurally similar to the References, so I put it after the
Discussion.


5. I reduced the explicit mention of the term "ESS-elsewhere".
After our discussion last week I realized that transparency to the user
is much more important for this paper than the technical detail.


git-svn-id: https://svn.r-project.org/ESS/trunk@1949 0bbaf3bd-34e0-0310-bf65-c717079852d4