1. change strings from *ddeclient ESS* to '(ddeESS [S+4])'
2. There is a weird error in this version that I don't understand.
The environment variable PATH gets the value of MANPATH.
I cannot figure out why.  It is MANPATH for me on two machines so far,
I don't know if that is related to my personal list of environment
variables or something else.  The fix I installed here works
on my machine, I hope it works on others.  I did attempt a generalization,
but that didn't work for me.


git-svn-id: https://svn.r-project.org/ESS/trunk@1954 0bbaf3bd-34e0-0310-bf65-c717079852d4