merge my changes with Tony's
Lots of little changes for smoothness.

SAS mode is currently defined in 3.3.SAS but is implicitly used in the
previous section {\SAS---Interactive in Emacs buffers.}
I didn't fix this.

Rodney, does sas-log-mode highlight warnings, lognotes, etc even when
the syntax of the sas language is turned off?  I think it would be
efficient.

Tony, I can live with the present organization or the revised ones you
proposed.


git-svn-id: https://svn.r-project.org/ESS/trunk@2016 0bbaf3bd-34e0-0310-bf65-c717079852d4