I added content and reformatted the history table. please check.
I still have questions about formatting the history table.  The
previous table had initials for the early SAS-mode authors, but not
for the early S-mode authors.  I removed all initials.  The full names
are in the text.  Since the history is here, I see no need for an
acknowledgments section.


Why is Mac listed for 1999, but not 2001?

many small changes.

question: why did someone add \centering to the figures?
I think it is redundant when the figures are forced to
\textwidth.  Am I missing something?


%%% Rich: section 3 has subsections, the other sections don't.
%%% We could promote the section 2 paragraphs to subsections, and
%%% the section 3 paragraphs to subsubsections.  My preference
%%% is to leave them as they are as I think all the units that are
%%% currently called paragraphs should have the same weight.
%%% I changed the phrasing of the cross-reference below.


git-svn-id: https://svn.r-project.org/ESS/trunk@2034 0bbaf3bd-34e0-0310-bf65-c717079852d4