commented out 2002 since this is "history".  removed request re: prev generations.
I would like to remove ESS Core names from the history as it pertains
to ESS (i.e. early 1997 was when we renamed the project), but leave in
external contributors at that point.  This is a style issue, and I'd
like comments on it.  I've not done this yet, but will today unless
people object.


git-svn-id: https://svn.r-project.org/ESS/trunk@2038 0bbaf3bd-34e0-0310-bf65-c717079852d4