S+elsewhere is the 1998 name, and belongs in the table.
ESS-elsewhere is the newer name, and the one that is still under
discussion.  When it settles down, I now think the name ess-remote is
a better name.

I switched the last history item to 2002 5.1.20.  Do we have any new content
over 19?  If so then we should have both listed.  The new items I think of
are S+6 and Emacs 21, both of which don't need to be mentioned.

SAS interactive was designed for remote usage.  I think that for local
usage it makes more sense to use emacs for editing and then cut and
paste to the SAS Edit window.

On names of ESS core members, something on the order of:
   In 1997 we renamed the project to ESS with Heiberger, Hornik,
   Maechler, and Rossini.  In 1998 Sparapani joined the core.
While I agree that assignment of specific topics to specific people is
not needed, it feels ok to me to have it there.  If we keep assignment
then Rossini should be identified as coordinator.


git-svn-id: https://svn.r-project.org/ESS/trunk@2055 0bbaf3bd-34e0-0310-bf65-c717079852d4