1. added black and white figures as a supplement.
2. revised the \section{ESS extends Emacs}

3. fixed sentences that used "which" twice within a few words of
   each other.

4. The recognition that \stexttt{iESS[S]} is what the user sees when
running M-x S+elsewhere is very slick.  It is correct.  The command
sequence to start it differs from the sequence on the local machine
(and correctly is too much detail for this paper), but once it is
functioning it is identical.

5. SAS Display Manager. one more attempt.

6. tweaked other little stuff.

7. merged 4.6 and 4.7 in the history and deleted reference to 4.6.

Files part of the paper to be sent to jcgs
  essJCGSv2.tex
  hilock-sas-bw.ps.bb
  ediff-sas-bw.ps.bb
  ess-demo-bw.ps.bb
  ess-demo-bw.ps.gz
  ediff-sas-bw.ps.gz
  hilock-sas-bw.ps.gz

Files for use by authors
  other.sas
  hilock-bw.sas
  ess-demo.s
  philasug.sas


git-svn-id: https://svn.r-project.org/ESS/trunk@2160 0bbaf3bd-34e0-0310-bf65-c717079852d4