The repair of the C-c C-l problem reported by S.McClatchie <s.mcclatchie@niwa.cri.nz> and Bj��rn-Helge Mevik <bhm@math.uio.no> required a change to ess-command in ess-iw32.el.  I believe the problem was limited to Windows machines and first appeared in 5.1.19 when I added ess-command to ess-iw32.el.
The temporary buffer buf was not associated with an ESS process.  Now it is.

This is a consequence of how I created a windows version of
ess-command.  Part of loading a file is checking whether the new file
would overwrite any old S language objects.  This in turn requires
running search() which uses ess-command.  The temporary buffer created
to hold the results of the search was not associated with an ESS
process.  Therefore when ess-command checks whether it is running
under Windows, it finds instead that it is in the tbuffer which is not
connected with an ESS process so it can't decide and quits.  Since the
original buffer that the user was trying to load was correctly
connected to an ESS process, the user is justifiably confused.

The repair is done at the most central point, at the ess-command for
windows.


git-svn-id: https://svn.r-project.org/ESS/trunk@2169 0bbaf3bd-34e0-0310-bf65-c717079852d4