I fixed up help-sas, installed help-s, fixed up some style issues, cleaned up cross references to empty pages.
All occurences of @essver MUST have braces @essver{} or else the remainder
of the source line is treated as a comment.

The doc/Makefile doesn't depend on the *.texi files.
I attempted to change it to
$(ESSINFODIR)/ess.info: *.texi
but that doesn't work.  So I used
$(ESSINFODIR)/ess.info: ess.texi help-sas.texi inst_tar.texi

The @display requires everything to be left justified.  Otherwise the
indentations are interpreted as hard spaces.

Why are some lines in @display and others in @example?  They look the same.
This is inst_tar.texi, lines 29-36
@example
	gunzip ess-@essver{}.tar.gz
	tar vxf ess-@essver{}.tar
@end example
@display
	(or: @code{gunzip < ess-@essver{}.tar.gz | tar vxf -} ).
	(or using GNU tar:  @code{tar zvxf ess-@essver{}.tar.gz}).
@end display

I added the control-function keys to help-sas.texi

I fixed the help menu pointers to the empty operating system nodes.

As long as I am doing info, I fixed up help-s.texi and installed it.
It compiles correctly.  It still needs to be read carefully to make sure
it is coherent.


I think the various doc/README* files now need to be regenerated
from the appropriate *.texi files.


git-svn-id: https://svn.r-project.org/ESS/trunk@2209 0bbaf3bd-34e0-0310-bf65-c717079852d4