S-Plus 6.1 beta for Windows works correctly with ESS. 6.0 didn't work correctly.
I modified M-x S+6 to
  "Verify that `inferior-S+6-program-name' points to S-Plus 6.
Start normally for S-Plus 6.1.  Inform the user to start S-Plus 6.0
from the icon and than connect to it with `S+6-existing'.  Give an error
message if `inferior-S+6-program-name' doesn't point to S-Plus 6."


S-Plus 6.1 starts up exceedingly slowly.  I changed the delay time in
M-x S+6-initiate to 60 seconds (double what we needed for M-x S+4).
The is the delay needed for my 300MHz machine.  I will make this a
variable so the users can customize it for their speed.


git-svn-id: https://svn.r-project.org/ESS/trunk@2220 0bbaf3bd-34e0-0310-bf65-c717079852d4