ess-remote
First draft of what I consider the right way to do what we
have previously called S+elsewhere and ESS-elsewhere.

Please test this.  Start a telnet or other protocol to a remote computer,
start S or R (or maybe something else), M-x ess-remote, and tell it which
dialect.  You are now talking to an inferior-ess process.

I added all the *-customize-alist to ess-select-alist-dialect.  I added
only sp6 to the dialects we prompt for, but all will be recognized.

Rodney, dialect "sas" points to S+elsewhere-customize-alist rather than
to SAS-customize-alist.

If this works then ess-add-ess-process should be moved from essa-sas.el to
ess-inf.el


git-svn-id: https://svn.r-project.org/ESS/trunk@2222 0bbaf3bd-34e0-0310-bf65-c717079852d4