S-Plus 6 help() function uses pager=options()$help.pager
I modified inferior-S-language-start in essl-s.el to always define
options()$help.pager to have the same value as options()$pager.
Only S-Plus 6 uses this new value.  The way the statement is constructed,
it does not get used for R and it generates a harmless extra component to
the options() in earlier versions of S-Plus.


git-svn-id: https://svn.r-project.org/ESS/trunk@2246 0bbaf3bd-34e0-0310-bf65-c717079852d4