Thank you Martin, I was starting do essentially the same thing to ess.texi.  Now I don't have to.
I did one extra step.  We had an empty chapter in the dvi that I commented out.
  @comment @chapter Help for Statistical Packages
Then I promoted both S and SAS to chapter status.
  @chapter Help for the S family
  @chapter Help for SAS


There is a fundamental problem in both help-s.texi and help-sas.texi.
Make the ess.dvi and then look at the S and SAS chapters.
Both chapters have a trailing piece of the node-name as the last line
in each section.  I tried several ways to make that go away with no
success.  Can you figure out what is going on and fix it?  Thanks


git-svn-id: https://svn.r-project.org/ESS/trunk@2249 0bbaf3bd-34e0-0310-bf65-c717079852d4