ESS[SAS]:  ess-sas-goto-shell resurrected!  Our recent discussion with respect to each SAS program, perhaps, needing it's own ess-sas-submit-method brought me back to ancient discussions of asynchronous shell buffer names.  Emacs and XEmacs named them differently.  Our solution was to use & for 'sh and start for 'ms-dos to make synchronous processes asynchronous.  However, this made ess-sas-goto-shell unnecessary since the buffer was always *shell*.
Now, I added the variable ess-sas-shell-buffer which is buffer-local
and defaults to *shell*.  If you set this to something else, then you
will get another shell buffer.  And, ess-sas-goto-shell will take you
to whichever one you need.  So, we are back to the original function.
I also moved the add-hook call to the shell creation step since it
seemed like over-kill to call it every time you did ess-sas-submit-sh.
Is this enough explanation or am I boring you?


git-svn-id: https://svn.r-project.org/ESS/trunk@2298 0bbaf3bd-34e0-0310-bf65-c717079852d4