Figure 1 is now satisfactory.  I changed the names of the revisions to fig1-top-bw* and fig1-bottom-bw*
exampl4.s is the file that is displayed in the fig1* files.

This was awful.  There must be a better way to do quality screen dumps.
I did alt-PrintScreen, then dropped each figure into Word.
In Word, I grabbed the corner of each figure and stretched it as far as it
would go (about 221% or something like that).
Then I saved the Word file in html mode, that produces a 200 byte html file
and two gif files.  I loaded each gif into xv, running in xfree86 on the PC,
and saved it as postscript grayscale in 11x17inch landscape.


git-svn-id: https://svn.r-project.org/ESS/trunk@2419 0bbaf3bd-34e0-0310-bf65-c717079852d4