ESS[SAS]: ess-sas-graph-view (F12) enhanced; you can specify external file viewers for each graphics file type via the alist ess-sas-graph-view-viewer-alist; also .jpg/.gif are now handled by image-mode on XEmacs, if available, otherwise by graphics primitives as before
New variables (the first two renamed, the last is new)
ess-sas-graph-view-suffix-regexp is a regular expression of supported
file types defined via file name extensions.
ess-sas-graph-view-viewer-default is the default external viewer for
your platform.  ess-sas-graph-view-viewer-alist is an alist of
exceptions to the default; i.e. file types and their associated
viewers which will be used rather than the default viewer.


git-svn-id: https://svn.r-project.org/ESS/trunk@2809 0bbaf3bd-34e0-0310-bf65-c717079852d4