This is the windows equivalent of ess-r-version-create. It is different because R has a different naming convention on windows.
A few questions before I integrate this and the ess-sqpe-versions-create.el
that I posted last week.

ess-site.el currently has
(ess-r-versions-create)
I hope this is fast since it is done every time.

ess-find-rterm is slow when it creates ess-rterm-versions and probably
should be done manually once per site when ESS is installed and when R
is upgraded.  I am not at all sure where in the documentation this
information should be placed.

ess-sqpe-versions needs to be created manually for each machine, but we can
give it some default values.

Where do these manually created variables go?  in ess-custom?

ess-sqpe-versions-create and ess-rterm-versions-create could be put into
ess-site.el.  Should they and ess-r-versions-create be conditioned on
ess-microsoft-p?  I think yes.

I want these on the menu.  Can someone write a menu statement for me
and condition it on a new ess-site variable?  The new variable would
be nil by default, but would be set to t by ess-core and r-core and
maybe some other users.


git-svn-id: https://svn.r-project.org/ESS/trunk@2876 0bbaf3bd-34e0-0310-bf65-c717079852d4