Several changes to   ess-sqpe-versions-create   ;; now in essdsp6w.el   ess-rterm-versions-create  ;; now in essd-r.el and a request for a change to   ess-r-versions-create
These functions are now integrated into
  ess-sqpe-versions-create
  ess-rterm-versions-create
essd-r.el and essdsp6w.el and ess-site.el and ess-cust.el

I removed the two files ess-find-rterm.el ess-sqpe-versions-create.el
that I posted earlier.

I rewrote ess-find-rterm so it now creates ess-rterm-versions fast.
ess-sqpe-versions still needs to be created manually if the defaults
aren't right.

I need help in placing these new functions on the menu.  I looked at
ess-mode.el and don't immediately see how to write something similar
when the number and names of the functions to be placed on the menu
are variable.  The names of the generated functions can be made to be
the explicit result of the generating functions (they aren't at the
moment, this is the change alluded to above).  Here is a mockup of
what I want it to look like:

(if ess-versions-on-menu  ;; this variable now exists
  (let* ((sqpe  (ess-sqpe-versions-create))
         (rterm (ess-rterm-versions-create))
         (r     (ess-r-versions-create)))
;; add to menu
("Start Other Versions"
 (map sqpe  ["splus61" splus61 ess-microsoft-p])       ;; I show the first
 (map rterm ["rw1090" rw1090   ess-microsoft-p])       ;; generated item from
 (map r     ["R-1.9.0" R-1.9.0 (not ess-microsoft-p)]) ;; each expansion.
)))

I defaulted ess-versions-on-menu to nil.  ESS-core and R-core, and maybe
others, would change that to t in their own .emacs files.

None of this new stuff is documented yet.


git-svn-id: https://svn.r-project.org/ESS/trunk@2878 0bbaf3bd-34e0-0310-bf65-c717079852d4