change the treatment of empty lines by ess-eval-line-and-step.
It skips empty lines before the evaluation and not after, as the old version
did. This is a more reasonable behavior when you execute a paragraph line by
line and repeatedly modify/execute the same portion of code. Previously this
function would just jump to the next paragraph or even worse, to the end of
buffer.

git-svn-id: https://svn.r-project.org/ESS/trunk@4915 0bbaf3bd-34e0-0310-bf65-c717079852d4