* ess-s-l.el (S-common-cust-alist): Change default for comment-use-syntax to t.  Dirk Eddelbuettel reported the follownig bug:
With a line like the following in an R buffer, when M-TAB is hit
on the line to add a comment, the hash within the string is
interpeted as an existing comment, causing the string to be
disrupted.

rug( foo,  col="#00000030")

This is now fixed such that M-TAB would add a hash at
the end of the line, allowing for a comment.

git-svn-id: https://svn.r-project.org/ESS/trunk@5609 0bbaf3bd-34e0-0310-bf65-c717079852d4