.log/ESS[LOG] auto-detection is broken.  .log's are erroneously detected as ESS[LST].  I took a look at the code, and I see the "The SAS System" trick is used for detecting .lst (as well as .log).
First of all, this will not work for .lst since "The SAS System"
rarely appears; most users turn off this annoyance.  But, second,
and more importantly, since the same trick is used for both, I 
suspect that is where the bug was introduced.

Since the release is imminent, I just checked in a simple workaround.
I split ess-automatic-sas-log-or-lst-mode into 
ess-automatic-sas-log-mode (default t) and ess-automatic-sas-lst-mode 
(default nil).  This seems to get around the issue for now.  But, we
need to re-visit this later.


git-svn-id: https://svn.r-project.org/ESS/trunk@5932 0bbaf3bd-34e0-0310-bf65-c717079852d4