Display knitr output buffers only on error
The patch to `ess-swv-run-in-R` will only display a new window when
there is an error in the R code and will display a message in the
minubuffer when the weaving/tangling/kniting/purling is complete.
`ess-swv-run-in-R` is currently only used for
weaving/tangling/kniting/purling, so this won't affect other aspects of
ESS.

The patch to `ess-swv-PDF` simply prevents the buffer with the TeX
output from being displayed when there are no errors compiling the TeX.
Instead, a status message is displayed in the minibuffer.

Motivation:

The usual routine to create a PDF file from a knitr Rnw document is to
run `ess-swv-knit` with `M-n r`, then `ess-swv-PDF` with `M-n P`.  It's
not hard to do `C-x 1` after these commands to close the newly displayed
window, but for anyone with multiple windows displaying, it can be
tedious to return to the previous window configuration.

TODO:

For consisency, update ess-swv-PS to mimic the behaviour of ess-swv-PDF.