Add ess-swv-weave-PDF to knit, compile, show PDF
The command `ess-swv-weave-PDF` will

1. weave or knit the noweb source document (depending on the value of
`ess-swv-processor`), only displaying the *R* buffer if there is an
error

2. compile the generated TeX document, only displaying
the *ESS-tex-output* buffer if there is an error

3. display the generated PDF document.

Optional arguments BLOCK and HIDE-COMPILE-BUFFER were added to the
functions `ess-swv-run-in-R` and `ess-swv-PDF`, respectively.  BLOCK
will cause `ess-swv-run-in-R` to wait for R to finish.  This is
necessary, because we need to wait for the .tex file to be generated
before attempting to compile it.  HIDE-COMPILE-BUFFER will prevent
the *ESS-tex-output* buffer from being displayed if it is hidden and
there are no errors.

`ess-swv-weave-PDF` is bound to M-n v ('v' for 'v'iew).

The behaviour of weaving/knitting and tangling/purling, changed in the
last commit, have been returned to the default behaviour, i.e.,
displaying the *R* and *ESS-tex-output* buffers.