Smarter package detection
Previously, only "R" directories were detected. Now, a whole ranges of
directories are detected. The challenge is to avoid recursive
algorithms because they are slow in remote environments.

The new system uses a new variable `ess-developer-package-dirs` which
contains an alist of directory names and their depth in the package
hierarchy. For example the "testthat" folder is usually at depth 2
compared to the package root. This can be user-customised.