Simplify ess-smart-S-assign and ess-smart-S-assign-key
This changes is so that users can (setq ess-smart-S-assign-key nil) at
any time to disable the smart assignment. To change which key is used,
they must set ess-smart-S-assign-key before ESS is loaded.

If ess-smart-S-assign-key is nil, ess-smart-S-assign behaves like
self-insert-command.

It keeps around the old functions so that they don't cause errors, but
they're ignored in favor of having users set ess-smart-S-assign-key
themselves.

Closes #558