Make ESS respect 'display-buffer-alist'
Now users can configure 'display-buffer-alist' to customize where and
how ESS's various buffers are displayed. This also deprecates the
existing ESS-specific framework we have in favor of offloading that
responsibility to Emacs. The docstrings of inferior-ess-same-window and
inferior-ess-own-frame were modified to include instructions on how to
take advantage of 'display-buffer-alist'.