Refactor ess-set-CRAN-mirror
This was originally reported on the mailing list. If you set
`ess-use-ido' to nil, we fall back on `completing-read'. Some
completing read engines (ido, ivy) do fine with text properties but
others (default, helm) seem to strip them. This refactors
ess-set-CRAN-mirror to not rely on text properties at all in order to
sidestep this issue.