Do not split help buffers when using own frame
* lisp/ess-inf.el (ess-eval-line-visibly-and-step)
  (ess-eval-region-or-line-visibly-and-step): Search all frames for the
  process buffer if ess-help-own-frame is non-nil.

Solves a bug first reported on the mail list; here is the original bug
report:

There's something strange going on with the R help buffers.  I'm using
the latest ess-20190122.2108 from melpa, but it's not particular to that
version:

I want each help buffer to appear in its own separate frame, and have

  '(ess-help-own-frame t)
  '(ess-help-reuse-window nil)

in my .emacs.  Now, pressing, say, '?lm' in an R buffer opens up the
help page in a separate frame as it should.  However, using 'l' to
execute the code in the 'Examples:' section results in the help frame
being split up with the help buffer on top and the R buffer below. Then,
pressing 'q' or 'k' should bury or remove the help frame, respectively,
but it's still visible and has the help buffer on top replaced by the R
buffer.
(ess-eval-region-or-line-visibly-and-step):
(ess-eval-line-visibly-and-step):