Fix roxy bug inserting parameters from previous function
Originally reported on the mailing list by Paul Johnson. With two
functions, 'ess-roxy-update-entry' would insert parameters from the
previous function. This is due to 'beginning-of-defun' getting called
with point already at the beginning of the function in
'ess-roxy-get-function-args'. Fix this by making
'ess-roxy-get-function-args' assume that point is at the beginning of
the function. Add tests.