Make 'ess-r-runners-prefixes' a defcustom
Make 'ess-r-versions' an obsolete alias of 'ess-r-runners-prefixes'.

This allows us to use the :set keyword, which allows us to regenerate
the runners if needed. New function 'ess-r-runners-reset' takes care
of that.

Update the manual with this new information, removing some
older information about S on Unix/Windows systems which is outdated.

Closes #857
See discussion in #858