Fix ess-roxy-update-entry when point is at beginning of a function
Thanks to Marius Hofort for this bug report via the ess-help mailing
list:

I realized that in the current snapshot version of ESS (18.10.3), C-c
C-o C-o (for Roxygen function headers) does not work as expected
anymore (on macOS 10.14.5) in the following case (and others):

fun1 <- function(x) x

fun2 <- function(x) x

If the point is right before the "f" (so column 0) of "fun2", then C-c
C-o C-o puts the Roxygen header above fun1, not fun2. If the point is
anywhere else on that line, the header is put in the right place, so
right above fun2.