Use output filter in `with-r-running`
Instead of ordinary filter which is for background tasks.

Because the ordinary filter (currently) does not track the process
mark, a test for visible region evaluation had to be updated. The
previous output was wrong because linewise eval inserts user inputs in
the process buffer, which interacted badly with the ordinary filter.
That's why there is now an additional `+`. Before this it was inserted
after the prompt (like this: `> +`) and (confusingly) removed from the
test output along with the prompt.