use different control codes as MPI start/end delimiters
Literal ESC is tricky because it can be consumed by the shell or otherwise
misinterpreted. Use FS to mark the beginning, GS to mark the end, and RS (which
was previously used) to mark the end of the header. Hopefully the separator
control codes are unlikely to cause conflicts.

Also simplify the delimiters to use *solely* the control code and not rely on
other adjacent ASCII characters.

Make corresponding updates to elisp variables
ess-mpi-message-{start,end}-delimiter and update the docstring for
`ess-mpi-handle-messages` (which curiously already used GS as the end
separator).