Remove trailing paren from regexp to find date from R --version
Since R 4.2.0, the windows version of R modified slightly the format
of the date output by `R --version`.  This minor change prevented
ess-r-version-date extracting the date.

e.g. the output (on Windows only; not mac/linux) looks like:

 R version 4.2.0 (2022-04-22 ucrt) -- "Vigorous Calisthenics"

If we remove the trailing ) in the regexp, the problem is solved.

(Another approach would be to search for the optional text 'ucrt'
but that seems like overkill for now.)